Move all the code into one package
