+++ requirements.txt
@@ -1,7 +1,9 @@
+Flask==1.1.1
 Jinja2==2.10.3
-PyGithub==1.44.1
+PyGithub==1.46
 docker==4.2.0
 pexpect==4.8.0
+pylev==1.3.0
 python-gnupg==0.4.5
 semver==2.9.0
 toml==0.10.0
