--- test/__init__.py
@@ -1,3 +1,0 @@
-import sys
-
-sys.path.insert(0, "../tagbot")
