+++ test/action/test_repo.py
@@ -9,8 +9,8 @@ import pytest
 
 from github import UnknownObjectException
 
-from tagbot import TAGBOT_WEB, Abort
-from tagbot.repo import Repo
+from tagbot.action import TAGBOT_WEB, Abort
+from tagbot.action.repo import Repo
 
 
 def _repo(
@@ -109,9 +109,9 @@ def test_commit_sha_of_tag():
     assert r._commit_sha_of_tag("v3.4.5") is None
 
 
-@patch("tagbot.repo.error")
-@patch("tagbot.repo.warn")
-@patch("tagbot.repo.info")
+@patch("tagbot.action.repo.error")
+@patch("tagbot.action.repo.warn")
+@patch("tagbot.action.repo.info")
 def test_filter_map_versions(info, warn, error):
     r = _repo()
     r._commit_sha_of_tree = Mock(return_value=None)
@@ -130,7 +130,7 @@ def test_filter_map_versions(info, warn, error):
     assert r._filter_map_versions({"4.5.6": "tree4"}) == {"v4.5.6": "sha"}
 
 
-@patch("tagbot.repo.debug")
+@patch("tagbot.action.repo.debug")
 def test_versions(debug):
     r = _repo()
     r._Repo__registry_path = "path"
@@ -174,7 +174,7 @@ def test_run_url():
         assert r._run_url() == "https://github.com/Foo/Bar/actions"
 
 
-@patch("tagbot.repo.warn")
+@patch("tagbot.action.repo.warn")
 @patch("docker.from_env")
 def test_image_id(from_env, warn):
     r = _repo()
@@ -213,7 +213,7 @@ def test_create_dispatch_event(post):
     )
 
 
-@patch("tagbot.repo.mkstemp", side_effect=[(0, "abc"), (0, "xyz")] * 3)
+@patch("tagbot.action.repo.mkstemp", side_effect=[(0, "abc"), (0, "xyz")] * 3)
 @patch("os.chmod")
 @patch("subprocess.run")
 @patch("pexpect.spawn")
@@ -263,8 +263,8 @@ def test_configure_ssh(spawn, run, chmod, mkstemp):
     spawn.return_value.assert_has_calls(calls)
 
 
-@patch("tagbot.repo.GPG")
-@patch("tagbot.repo.mkdtemp", return_value="gpgdir")
+@patch("tagbot.action.repo.GPG")
+@patch("tagbot.action.repo.mkdtemp", return_value="gpgdir")
 @patch("os.chmod")
 def test_configure_gpg(chmod, mkdtemp, GPG):
     r = _repo()
