--- web/.gitignore
@@ -1,3 +1,0 @@
-.requirements/
-.serverless/
-node_modules/
