--- web/requirements.txt
@@ -1,3 +1,0 @@
-PyGithub==1.46
-Flask==1.1.1
-pylev==1.3.0
